Scripts to generate initial condition improved
The new class to minimize the energy functional in spherical harmonic basis can now be used to generate initial condition. In both scripts to generate initial condition there is an extra optional argument to fix the seed of the random number generator, and as such, we can generate the same pseudo-random state as many times as we want just by using the same seed.