New utilities
Added function to compute first and second order derivatives for both theta and phi in the entire grid with a 4-th order scheme. Introduced function to compute average residue in the grid. New data structure introduce to hold in the same place the Complex array correponding to states of species A and B simultaneously. Also provide arrays to hold the real and imaginary parts.